+++ tests/src/Unit/Resolver/MinutelyResolverTest.php
@@ -13,6 +13,7 @@ use Drupal\Tests\timesup\Traits\InvokeMethodTrait;
 use Drupal\Tests\UnitTestCase;
 use Drupal\timesup\Periodicity\MinutelyResolver;
 use PHPUnit\Framework\Attributes\AllowMockObjectsWithoutExpectations;
+use PHPUnit\Framework\Attributes\DataProvider;
 
 /**
  * @coversDefaultClass \Drupal\timesup\Periodicity\MinutelyResolver
@@ -141,6 +142,7 @@ class MinutelyResolverTest extends UnitTestCase {
    *
    * @dataProvider shouldApplyProvider
    */
+  #[DataProvider('shouldApplyProvider')]
   public function testShouldApply($request_time, $last_run, $expected) {
     $this->settings->expects($this->once())
       ->method('get')->with('resolvers')->willReturn(['minutely' => TRUE]);